[KMSv2] restructure kms staging dir
